pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["volochay <evgenia@volochay.info>"]
readme = "README.md"
packages = [
  { include = "task_manager" },
]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.7"
dj-database-url = "^0.5.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.5"
gunicorn = "^20.1.0"
django-bootstrap4 = "^22.3"
flake8 = "^4.0.1"
black = "^23.3.0" 
model-bakery = "^1.10.3"
django-filter = "^23.1"
pytest = "^7.2.2"
pytest-cov = "^4.0.0"
coverage = "^7.2.3"
codeclimate-test-reporter = "^0.2.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
